+++ NutriTechBackOffice/ClientApp/src/app/pages/listado-usuarios/listado-usuarios.component.ts
@@ -3,6 +3,7 @@ import { MatDialog } from "@angular/material";
 import { MatPaginator } from "@angular/material/paginator";
 import { MatSort } from "@angular/material/sort";
 import { MatTableDataSource } from "@angular/material/table";
+import { PopEliminarComponent } from "src/app/components/pop-eliminar/pop-eliminar.component";
 import { AuthService } from "src/app/services/auth.service";
 import { PopUpComponent } from "../../components/pop-up/pop-up.component";
 import { User } from "../../interfaces/user";
@@ -81,11 +82,11 @@ export class ListadoUsuariosComponent implements OnInit {
       console.log(this.userClaims);
       if (this.userClaims != null) {
         if (this.userClaims.Nutricionista) {
-          console.log("es nutricionista")
+          console.log("es nutricionista");
           this.loadNutrittionistGrid();
         }
         if (this.userClaims.Admin) {
-          console.log("es admin")
+          console.log("es admin");
           this.loadAdminGrid();
         }
       }
@@ -106,19 +107,33 @@ export class ListadoUsuariosComponent implements OnInit {
     }
   }
 
-  eliminar(email: string) {
-    this.usersService.deleteUser(email).subscribe(
-      (data) => {
-        this.dialog.open(PopUpComponent, {
-          data: { title: "Listo!", message: "El usuario fue eliminado." },
-        });
-        this.cargarGrilla();
-      },
-      (err) => {
-        this.dialog.open(PopUpComponent, {
-          data: { title: "Oops!", message: "El usuario no se pudo eliminar." },
-        });
+  openDialog(email: string) {
+    const dialogRef = this.dialog.open(PopEliminarComponent, {
+      data: {
+        message: 'Esta seguro que quiere eliminar?',
+        buttonText: {
+          ok: 'Si',
+          cancel: 'No'
+        }
       }
-    );
+    });
+
+    dialogRef.afterClosed().subscribe((confirmed: boolean) => {
+      if (confirmed) {
+        this.usersService.deleteUser(email).subscribe(
+          (data) => {
+            this.dialog.open(PopUpComponent, {
+              data: { title: "Listo!", message: "El usuario fue eliminado." },
+            });
+            this.cargarGrilla();
+          },
+          (err) => {
+            this.dialog.open(PopUpComponent, {
+              data: { title: "Oops!", message: "El usuario no se pudo eliminar." },
+            });
+          }
+        );
+      }
+    });
   }
 }
